generated_endpoints_router and neuroagent_endpoints_router for fastapi. Basic hello world example for the latter
